requirements.txt
# Bleeding edge Django
django==1.11.17

# Configuration
django-cache-url==2.0.0
dj-database-url==0.4.2

# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1
django-floppyforms==1.7.0

# Models
django-model-utils==3.0.0

# images
Pillow==5.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.32.0

# For the persistance stores
psycopg2==2.8.3


# Useful things
django-avatar==3.1.0

# Your custom requirements go here

# OAUTH2 support library
twilio==6.1.1

django-haystack==2.7.dev1 

whoosh==2.7.4
django-postman==3.5.1
django-taggit==0.22.1
django-formtools==2.0
django-localflavor==1.4.1
gunicorn==19.7.1
django-mailgun==0.9.1
django-versatileimagefield==1.10
django-debug-toolbar==1.8.0
django-storages==1.5.2
boto3==1.4.4
django-permanent==1.1.6
django-widget-tweaks==1.4.1